pyproject.toml

[project]
name = "vmPipeline"
version = "0.0.1"
requires-python = ">=3.9.4"
dependencies = [
    "google-cloud-compute",
    "google-cloud-pubsub",
    "google-cloud-storage",
    "google-cloud-logging",
    "typer",
    "pydantic"
]

[project.optional-dependencies]
dev = [
    "pytest ~= 7.2.1",
    "mypy ~= 1.0.1",
    "black"
]


[project.scripts]
ml_pipeline = "app:main"
